requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
jinja2==3.1.4
python-multipart==0.0.9
structlog==24.4.0

psycopg[binary]==3.2.3
alembic==1.13.3
email-validator==2.2.0
python-dateutil==2.9.0.post0

httpx==0.27.2         # Slack webhook / HTTP cliente
APScheduler==3.10.4   # ya lo usas en main.py para el scheduler
beautifulsoup4==4.12.3
lxml==5.3.0           # parser para BeautifulSoup("lxml")
pypdf==4.3.1          # extracción de texto de PDFs
xmlschema==3.4.2
tenacity==8.3.0

#ia
scikit-learn==1.5.2
joblib==1.4.2

reportlab==4.2.2